+++ is_valid/test.py
@@ -1,26 +1,20 @@
 from .to_pred import to_pred
 
 
-class assert_valid(object):
+def assert_valid(data, predicate, context={}, message=None, advanced=True):
     """
-    Creates a function that asserts that the data is valid according to the
-    given predicate. If no ``message`` is provided to this function the
-    explanation of the predicate will be used for the AssertionError in case
-    the assertion fails.
+    Asserts that the data is valid according to the given predicate. If no
+    ``message`` is provided to this function the explanation of the predicate
+    will be used for the AssertionError in case the assertion fails.
     """
+    predicate = to_pred(predicate)
 
-    def __init__(self, predicate, context={}):
-        self._predicate = to_pred(predicate)
-        self._context = context
+    if message is None:
+        explanation = predicate.explain(data, context=context)
+        valid = explanation.valid
+        message = (repr if advanced else str)(explanation)
+    else:
+        valid = predicate(data)
 
-    def __call__(self, data, context={}, message=None, advanced=True):
-        if message is None:
-            c = self._context.copy()
-            c.update(context)
-            explanation = self._predicate.explain(data, context=c)
-            valid = explanation.valid
-            message = (repr if advanced else str)(explanation)
-        else:
-            valid = self._predicate(data)
-        if not valid:
-            raise AssertionError(message)
+    if not valid:
+        raise AssertionError(message)